This is probably the big(gest) architectural change for the next version.
TJShow:
- Added instancing. Using instancing, a timeline may be 'instantiated' multiple times. For each timeline, there always exists a 'static' instance (like in all previous versions of TJShow). Timelines can contain an 'InstancerTrack', which will create an instance when the timeline starts playing back. This instancer track can send triggers to the instanced timeline and can set initial values for some variables in the instanced timeline. The instanced timeline works like a normal timeline, except that there is no timeline window or live controls. When an instanced timeline triggers another timeline, it triggers the static instance of that timeline (when a cue in the instanced timeline triggers another cue in that instanced timeline, it is triggered inside the instance of course). An instance can have its own InstancerTracks. Each instance has its own copy of the local variables of the timeline (and Timeline::GetLocalVariables returns the local variables of the static instance; this is also the version that instances make a copy of).

Lots of changes were also made in the way outlet changes are propagated. Although outlet connections are set only once per track, changes from different players will result in different variable changes. To accomodate for this, players now cannot call outlet->Set(...) anymore, but instead have to use Playback::SetOutletValue. 

Outlet values over the network are also supported. Because channel numbers are now dynamically allocated, all the server has to do is find the right instance and track for that channel/group combination and set the outlet. 

Because of the instancing functionality, lots of things have been refactored. The most important change is that the Model class now only holds the timelines and has no connections to controllers or instances. The Timeline class now also holds local variables (although saving it, like the cuelist, is done externally; this will be fixed in the future, but it will cause the file format to change slightly). A new Instances class manages running instances and will create the 'root instance' (from which all other instances are descendants). Controller is abstracted into an 'Instance' interface. This is a change to accomodate for different Controller types in the future (to test implementations that use worker threads, for example). 

- The check for whether there are any clients addressed for a track that is about to be played (in Controller:SetPlaybackState) now doesn't complain anymore when a track is addressed to the 'all clients' group.

TJNP/TJShow:
- Added a new ActionResetChannel message which allows the server to reset a single channel. This makes usage of dynamically allocated channels much safer, since the same channel number may be used by different plugin types consequently. 

git-svn-id: https://svn.pixelspark.nl/svn/corespark/trunk@458 c9a305e2-9f8a-4be3-8c95-b0e5bab41dc9